+++ source/soccer/assignment/cost_gk.cpp
@@ -2,9 +2,9 @@
 
 namespace Tyr::Soccer
 {
-float Ai::gkRoleCost(const int t_robot_idx, const Assignment &t_assignment) const
+int Ai::gkRoleCost(const int t_robot_idx, const Assignment &t_assignment) const
 {
     (void) t_assignment;
-    return m_ref_state.ourInfo().gk_id == t_robot_idx ? 0.0f : -1.0f;
+    return m_ref_state.ourInfo().gk_id == t_robot_idx ? 0 : -1;
 }
 } // namespace Tyr::Soccer
